Bunker destruction effect
Added the effect of destruction for Bunkers. The script was taken from the tutorial (it is complex to understand).

Added player's score and lives on the screen with their corresponding operation.

Now the Game over menu will appear when the Player loses all his lives and the level can be restarted.